+++ kernel/src/memory/frame_allocator.rs
@@ -1,3 +1,7 @@
+//! Frame allocators for use in allocation and deallocation
+//! Contains a GlobalFrameAllocator, which is a wrapper around
+//! the BootIntoFrameAllocator and the BitmapFrameAllocator
+
 use crate::memory::{
     bitmap_frame_allocator::BitmapFrameAllocator, boot_frame_allocator::BootIntoFrameAllocator,
 };
@@ -16,6 +20,11 @@ pub enum GlobalFrameAllocator {
 }
 
 unsafe impl FrameAllocator<Size4KiB> for GlobalFrameAllocator {
+    /// Global allocator that allocates a frame using either the boot frame allocator or the bitmap
+    /// depending on what the current selected allocator is
+    ///
+    /// # Returns
+    /// The allocated frame
     fn allocate_frame(&mut self) -> Option<PhysFrame<Size4KiB>> {
         match self {
             GlobalFrameAllocator::Boot(ref mut boot_alloc) => boot_alloc.allocate_frame(),
@@ -25,6 +34,14 @@ unsafe impl FrameAllocator<Size4KiB> for GlobalFrameAllocator {
 }
 
 impl FrameDeallocator<Size4KiB> for GlobalFrameAllocator {
+    /// Global deallocator that calls either the boot frame allocator or the bitmap
+    /// depending on what the current selected allocator is
+    ///
+    /// # Arguments
+    /// * `frame`: the frame to deallocate
+    ///
+    /// # Safety
+    /// The frame must be ensured to be allocated
     unsafe fn deallocate_frame(&mut self, frame: PhysFrame<Size4KiB>) {
         match self {
             GlobalFrameAllocator::Boot(ref mut boot_alloc) => boot_alloc.deallocate_frame(frame),
@@ -36,15 +53,30 @@ impl FrameDeallocator<Size4KiB> for GlobalFrameAllocator {
 }
 
 /// Exposed function to allocate a frame that runs the global's allocate_frame
+///
+/// # Returns
+/// The allocated frame
 pub fn alloc_frame() -> Option<PhysFrame> {
     with_generic_allocator(|allocator| allocator.allocate_frame())
 }
 
-/// Exposed function to allocate a frame that runs the global's deallocate_frame
+/// Exposed function to deallocate a frame that runs the global's deallocate_frame
+///
+/// # Arguments
+/// * `frame`: The frame to deallocate
 pub fn dealloc_frame(frame: PhysFrame<Size4KiB>) {
     with_generic_allocator(|allocator| unsafe { allocator.deallocate_frame(frame) })
 }
 
+/// Gives access to the bitmap frame allocator to any passed in closure
+/// Example:
+/// with_bitmap_frame_allocator(|allocator| {
+///     // code to run
+/// })
+///
+/// Arguments:
+///
+/// * `f`: The closure to run
 pub fn with_bitmap_frame_allocator<F, R>(f: F) -> R
 where
     F: FnOnce(&mut BitmapFrameAllocator) -> R,
@@ -57,6 +89,15 @@ where
     f(alloc)
 }
 
+/// Gives access to the boot into frame allocator to any passed in closure
+/// Example:
+/// with_boot_into_frame_allocator(|allocator| {
+///     // code to run
+/// })
+///
+/// Arguments:
+///
+/// * `f`: The closure to run
 pub fn with_boot_into_frame_allocator<F, R>(f: F) -> R
 where
     F: FnOnce(&mut BootIntoFrameAllocator) -> R,
@@ -69,6 +110,15 @@ where
     f(alloc)
 }
 
+/// Gives access to the global frame allocator to any passed in closure
+/// Example:
+/// with_generic_allocator(|allocator| {
+///     // code to run
+/// })
+///
+/// Arguments:
+///
+/// * `f`: The closure to run
 pub fn with_generic_allocator<F, R>(f: F) -> R
 where
     F: FnOnce(&mut GlobalFrameAllocator) -> R,
