+++ kernel/src/memory/paging.rs
@@ -4,8 +4,7 @@
 
 use x86_64::{
     structures::paging::{
-        FrameDeallocator, Mapper, OffsetPageTable, Page, PageTable, PageTableFlags, PhysFrame,
-        Size4KiB,
+        Mapper, OffsetPageTable, Page, PageTable, PageTableFlags, PhysFrame, Size4KiB,
     },
     VirtAddr,
 };
@@ -16,13 +15,9 @@ use crate::{
         frame_allocator::{alloc_frame, dealloc_frame, FRAME_ALLOCATOR},
         tlb::tlb_shootdown,
     },
-    processes::process::PCB,
 };
 
-use super::{
-    frame_allocator::{with_bitmap_frame_allocator, with_generic_allocator},
-    HHDM_OFFSET,
-};
+use super::HHDM_OFFSET;
 
 static mut NEXT_EPH_OFFSET: u64 = 0;
 
@@ -203,7 +198,7 @@ pub fn map_kernel_frame(
 ///
 /// # Safety
 ///
-/// TODO
+/// Updating the flags for a page may result in undefined behavior
 pub fn update_permissions(page: Page, mapper: &mut impl Mapper<Size4KiB>, flags: PageTableFlags) {
     let _ = unsafe {
         mapper
@@ -423,65 +418,4 @@ mod tests {
         let mut mapper = MAPPER.lock();
         remove_mapped_frame(page, &mut *mapper);
     }
-}
-
-/// Clear the PML4 associated with the PCB
-///
-/// * `pcb`: The process PCB to clear memory for
-pub fn clear_process_frames(pcb: &mut PCB) {
-    let pml4_frame = pcb.pml4_frame;
-    let mapper = unsafe { pcb.create_mapper() };
-
-    with_generic_allocator(|deallocator| {
-        // Iterate over first 256 entries (user space)
-        for i in 0..256 {
-            let entry = &mapper.level_4_table()[i];
-            if entry.is_unused() {
-                continue;
-            }
-
-            let pdpt_frame = PhysFrame::containing_address(entry.addr());
-            unsafe {
-                free_page_table(pdpt_frame, 3, deallocator, HHDM_OFFSET.as_u64());
-            }
-        }
-        unsafe { deallocator.deallocate_frame(pml4_frame) };
-    });
-
-    with_bitmap_frame_allocator(|alloc| {
-        alloc.print_bitmap_free_frames();
-    });
-}
-
-/// Helper function to recursively multi level page tables
-///
-/// * `frame`: the current page table frame iterating over
-/// * `level`: the current level of the page table we're on
-/// * `deallocator`:
-/// * `hhdm_offset`:
-unsafe fn free_page_table(
-    frame: PhysFrame,
-    level: u8,
-    deallocator: &mut impl FrameDeallocator<Size4KiB>,
-    hhdm_offset: u64,
-) {
-    let virt = hhdm_offset + frame.start_address().as_u64();
-    let table = unsafe { &mut *(virt as *mut PageTable) };
-
-    for entry in table.iter_mut() {
-        if entry.is_unused() {
-            continue;
-        }
-
-        if level > 1 {
-            let child_frame = PhysFrame::containing_address(entry.addr());
-            free_page_table(child_frame, level - 1, deallocator, hhdm_offset);
-        } else {
-            // Free level one page
-            let page_frame = PhysFrame::containing_address(entry.addr());
-            deallocator.deallocate_frame(page_frame);
-        }
-        entry.set_unused();
-    }
-    deallocator.deallocate_frame(frame);
 }
