+++ kernel/src/memory/tlb.rs
@@ -1,12 +1,18 @@
-use core::arch::asm;
-
-use x86_64::VirtAddr;
+//! Translation Lookaside Buffer Shootdowns
+//!
+//! - Exposes a function to perform TLB Shootdowns
 
 use crate::{
     constants::{idt::TLB_SHOOTDOWN_VECTOR, MAX_CORES},
     interrupts::x2apic::{current_core_id, send_ipi, TLB_SHOOTDOWN_ADDR},
 };
+use core::arch::asm;
+use x86_64::VirtAddr;
 
+/// Sends an inter-process interrupt to all other cores to clear TLB entry with a specific VA
+///
+/// # Arguments:
+/// * target_vaddr: VA that has to be flushed in all TLBs
 pub fn tlb_shootdown(target_vaddr: VirtAddr) {
     let current_core = current_core_id();
     let vaddr = target_vaddr.as_u64();
